DynamoConnect now only in Pipeline folder
